+++ src/containers/Form/OutboundIdentifierLists/index.js
@@ -8,7 +8,6 @@ import { reduxForm } from 'redux-form/immutable';
 import OutboundIdentifierListsForm from './layout';
 import { formValidation } from './validation';
 import { getSelectedEntityId, isCreating } from '../../../redux/modules/entities/selectors';
-import { selectFlowIds } from '../../../redux/modules/entities/flows/selectors';
 import { selectFormInitialValues, formSubmission, createFormName } from '../../../redux/modules/form/selectors';
 
 const CreateOutboundIdentifierListsForm = compose(
@@ -22,7 +21,6 @@ const CreateOutboundIdentifierListsForm = compose(
 
 export function mapStateToProps(state) {
   return {
-    flowIds: selectFlowIds(state),
     initialValues: selectFormInitialValues(state),
     isSaving: isCreating(state),
     key: getSelectedEntityId(state)