+++ src/containers/Form/Users/index.js
@@ -13,15 +13,21 @@ import {
   isCreating,
   userHasUpdatePermission
 } from '../../../redux/modules/entities/selectors';
-import { selectFormInitialValues, formSubmission, createFormName } from '../../../redux/modules/form/selectors';
+import { onFormSubmit } from '../../../redux/modules/entities';
+import { selectFormInitialValues, createFormName } from '../../../redux/modules/form/selectors';
 import { getUserTenantStatus, getInvitationScenario } from '../../../redux/modules/entities/users/selectors';
 import { selectTenantRoles, selectPlatformRoles } from '../../../redux/modules/entities/roles/selectors';
+import { selectTenantIdentityProviders } from '../../../redux/modules/entities/identityProviders/selectors';
 
 const CreateUsersForm = compose(
   connect(createFormName),
   reduxForm({
-    onSubmit: formSubmission,
+    onSubmit: (values, dispatch, props) => {
+      return dispatch(onFormSubmit(values, props));
+    },
     validate: formValidation,
+    enableReinitialize: true,
+    keepDirtyOnReinitialize: true,
     destroyOnUnmount: true
   })
 )(UsersForm);
@@ -30,6 +36,7 @@ export function mapStateToProps(state) {
   return {
     platformRoles: selectPlatformRoles(state),
     tenantRoles: selectTenantRoles(state),
+    tenantIdentityProviders: selectTenantIdentityProviders(state),
     status: getUserTenantStatus(state),
     scenario: getInvitationScenario(state),
     initialValues: selectFormInitialValues(state),