+++ src/containers/Form/Users/layout.js
@@ -11,7 +11,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import { DetailHeader, InputField, SelectField, ToggleField } from 'cx-ui-components';
+import { DetailHeader, InputField, SelectField } from 'cx-ui-components';
 import DetailWrapper from '../../../components/DetailWrapper';
 
 const Wrapper = styled.div`
@@ -29,8 +29,8 @@ export default function UsersForm({
   tenantRoles,
   status,
   scenario,
+  tenantIdentityProviders,
   isSaving,
-  inherited,
   userHasUpdatePermission,
   key
 }) {
@@ -38,14 +38,14 @@ export default function UsersForm({
     <form onSubmit={handleSubmit} key={key}>
       <Wrapper>
         <DetailWrapper open={true}>
-          <WrappedDetailHeader text="Details" />
+          <WrappedDetailHeader text="Platform Details" />
           <InputField
             name="firstName"
             label="First Name"
             id="frm-users-first-name"
             componentType="input"
             inputType="text"
-            disabled={isSaving || inherited || !userHasUpdatePermission}
+            disabled={isSaving || !userHasUpdatePermission}
           />
           <InputField
             name="lastName"
@@ -53,7 +53,7 @@ export default function UsersForm({
             id="frm-users-last-name"
             componentType="input"
             inputType="text"
-            disabled={isSaving || inherited || !userHasUpdatePermission}
+            disabled={isSaving || !userHasUpdatePermission}
           />
           <InputField
             name="externalId"
@@ -61,45 +61,58 @@ export default function UsersForm({
             id="frm-users-external-id"
             componentType="input"
             inputType="text"
-            disabled={isSaving || inherited || !userHasUpdatePermission}
+            disabled={isSaving || !userHasUpdatePermission}
           />
+        </DetailWrapper>
+
+        <DetailWrapper open={true}>
+          <WrappedDetailHeader text="Tenant Details" />
           <InputField
             name="workStationId"
             label="Workstation ID"
             id="frm-users-work-station-id"
             componentType="input"
             inputType="text"
-            disabled={isSaving || inherited || !userHasUpdatePermission}
+            disabled={isSaving || !userHasUpdatePermission}
+          />
+          <SelectField
+            name="roleId"
+            label="Tenant Role"
+            id="frm-users-role-id"
+            disabled={isSaving || !userHasUpdatePermission}
+            options={tenantRoles}
           />
         </DetailWrapper>
 
         <DetailWrapper open={true}>
-          <WrappedDetailHeader text="Login" />
+          <WrappedDetailHeader text="Login Details" />
           <InputField name="email" label="Email" id="frm-users-email" componentType="input" inputType="text" disabled />
-          <SelectField
-            name="platformRoleId"
-            label="Platform Role *"
-            id="frm-users-platform-role-id"
-            disabled={isSaving || inherited || !userHasUpdatePermission}
-            options={platformRoles}
-            required
+          <InputField
+            name="platformStatus"
+            label="Platform Status"
+            id="frm-users-platform-status"
+            componentType="input"
+            inputType="text"
+            disabled
           />
           <SelectField
-            name="roleId"
-            label="Tenant Role *"
-            id="frm-users-role-id"
-            disabled={isSaving || inherited || !userHasUpdatePermission}
-            options={tenantRoles}
-            required
+            name="noPassword"
+            label="Platform Authentication"
+            id="frm-users-no-password-id"
+            disabled={isSaving || !userHasUpdatePermission}
+            options={[
+              { label: 'Use Tenant Default: Enabled', value: 'null' },
+              { label: 'Enabled', value: true },
+              { label: 'Disabled', value: false }
+            ]}
           />
-          {status !== 'invited' && (
-            <ToggleField
-              name="inviteNow"
-              label="Invite Now"
-              id="frm-users-invite-now"
-              disabled={isSaving || inherited || !userHasUpdatePermission}
-            />
-          )}
+          <SelectField
+            name="defaultIdentityProvider"
+            label="Single Sign On Identitiy Provider"
+            id="frm-users-default-identity-provider-id"
+            disabled={isSaving || !userHasUpdatePermission}
+            options={tenantIdentityProviders}
+          />
         </DetailWrapper>
       </Wrapper>
     </form>
@@ -116,6 +129,12 @@ UsersForm.propTypes = {
     PropTypes.shape({
       label: PropTypes.string,
       value: PropTypes.string.isRequired
+    })
+  ),
+  tenantIdentityProviders: PropTypes.arrayOf(
+    PropTypes.shape({
+      label: PropTypes.string,
+      value: PropTypes.string
     })
   ),
   tenantRoles: PropTypes.arrayOf(