+++ src/containers/Form/Users/validation.js
@@ -5,7 +5,7 @@
 import { isEmpty } from 'serenova-js-utils/strings';
 
 export const formValidation = values => ({
-  email: isEmpty(values.get('email')) && 'Please enter an email',
-  platformRoleId: !values.get('platformRoleId') && 'Please select a platform role',
-  roleId: !values.get('roleId') && 'Please select a tenant role'
+  // email: isEmpty(values.get('email')) && 'Please enter an email',
+  // platformRoleId: !values.get('platformRoleId') && 'Please select a platform role',
+  // roleId: !values.get('roleId') && 'Please select a tenant role'
 });