+++ src/redux/modules/entities/metaData.js
@@ -128,27 +128,28 @@ export class EntityMetaData {
 export const listOfEntities = [
   'branding',
   'chatWidgets',
-  'protectedBranding',
-  'flows',
-  'listTypes',
-  'lists',
+  'customMetrics',
+  'dashboards',
+  'dataAccessReports',
   'emailTemplates',
+  'flows',
+  'groups',
+  'historicalReportFolders',
+  'identityProviders',
+  'interactionMonitoring',
+  'lists',
+  'listTypes',
+  'messageTemplates',
   'outboundIdentifiers',
   'outboundIdentifierLists',
-  'customMetrics',
-  'groups',
-  'skills',
-  'roles',
-  'platformRoles',
   'permissions',
-  'interactionMonitoring',
-  'users',
+  'platformRoles',
+  'protectedBranding',
   'reasonLists',
-  'messageTemplates',
+  'roles',
+  'skills',
   'transferLists',
-  'historicalReportFolders',
-  'dashboards',
-  'dataAccessReports'
+  'users'
   //Hygen-insert-at-end-of-list
 ];
 
@@ -158,8 +159,8 @@ listOfEntities.forEach(x => (entities[x] = new EntityMetaData(x)));
 // Users
 entities.users.pageTitle = 'User Management';
 entities.users.helpLink = '/Help/Content/Managing%20Users/Adding_users.htm';
-entities.users.createFormDependencies.push('roles', 'platformRoles');
-entities.users.updateFormDependencies.push('roles', 'platformRoles');
+entities.users.createFormDependencies.push('roles', 'platformRoles', 'identityProviders');
+entities.users.updateFormDependencies.push('roles', 'platformRoles', 'identityProviders');
 entities.users.memberListTableFields = [
   { label: 'First Name', name: 'firstName' },
   { label: 'Last Name', name: 'lastName' },